+++ docs/court_wizard.js
@@ -256,24 +256,24 @@ if (!('encodeInto' in cachedTextEncoder)) {
 
 let WASM_VECTOR_LEN = 0;
 
-function __wasm_bindgen_func_elem_125929(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_125929(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_126061(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_126061(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_125930(arg0, arg1, arg2, arg3) {
-    wasm.__wasm_bindgen_func_elem_125930(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
+function __wasm_bindgen_func_elem_121531(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_121531(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_124369(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_124369(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_124501(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_124501(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_125928(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_125928(arg0, arg1);
+function __wasm_bindgen_func_elem_126060(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_126060(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_121399(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_121399(arg0, arg1);
+function __wasm_bindgen_func_elem_126062(arg0, arg1, arg2, arg3) {
+    wasm.__wasm_bindgen_func_elem_126062(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
 }
 
 const __wbindgen_enum_GpuAddressMode = ["clamp-to-edge", "repeat", "mirror-repeat"];
@@ -2222,14 +2222,19 @@ function __wbg_get_imports() {
         const ret = getObject(arg0).y;
         return ret;
     };
-    imports.wbg.__wbindgen_cast_00a01f595de2c16c = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("Event")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
+    imports.wbg.__wbindgen_cast_003d1f05065ec035 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_21a844c6344ca552 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 82841, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125930);
+    imports.wbg.__wbindgen_cast_0c6594e1d2a3e50e = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_1bb806bb300ff921 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("Event")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_2241b6af4c4b2941 = function(arg0, arg1) {
@@ -2237,39 +2242,39 @@ function __wbg_get_imports() {
         const ret = getStringFromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_236b428a67a2efdb = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
+    imports.wbg.__wbindgen_cast_54f16a8a1a0c9268 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_322b6f8c893904a0 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
+    imports.wbg.__wbindgen_cast_566e0ded1e801ea4 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82885, function: Function { arguments: [Externref], shim_idx: 82886, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_124491, __wasm_bindgen_func_elem_124501);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_4e3d2f1f67284add = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
+    imports.wbg.__wbindgen_cast_5f6c39be44b508c3 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_5792b1f8c02651d3 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
+    imports.wbg.__wbindgen_cast_6fb8c6e88310db27 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [], shim_idx: 82945, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126060);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_5e082aac6322f1f4 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
+    imports.wbg.__wbindgen_cast_9c0de4a26f099bf9 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_6d61e4e7eaad339d = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82780, function: Function { arguments: [Externref], shim_idx: 82781, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_124359, __wasm_bindgen_func_elem_124369);
+    imports.wbg.__wbindgen_cast_afcd701591a633a3 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 82946, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126062);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_7bb4b281101bfa14 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82212, function: Function { arguments: [], shim_idx: 82213, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_121378, __wasm_bindgen_func_elem_121399);
+    imports.wbg.__wbindgen_cast_b4f264ac3c065bcc = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82943, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 82944, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125939, __wasm_bindgen_func_elem_126061);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_cb9088102bce6b30 = function(arg0, arg1) {
@@ -2282,14 +2287,9 @@ function __wbg_get_imports() {
         const ret = arg0;
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_e045692bdf83284e = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 82839, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125929);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_e538d56e65352617 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 82838, function: Function { arguments: [], shim_idx: 82840, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_125807, __wasm_bindgen_func_elem_125928);
+    imports.wbg.__wbindgen_cast_eceb86d2f099ffb3 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 82317, function: Function { arguments: [], shim_idx: 82318, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_121510, __wasm_bindgen_func_elem_121531);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_object_clone_ref = function(arg0) {